CA-78802: Fix bug returning from signal
If eax is -513 (-ERESTARTNOINTR, see handle_signal in signal.c), you get a
signal and inside that signal in the kernel you have a xen callback the
callback is handled as a syscall corrupting the eip.
This patch set orig_ax in pt_regs so the xen handler is not handled as a
syscall but as an interrupt.