CP-3759: add ABI version to EXTRAVERSION
Add the in-development ABI version (0) to EXTRAVERSION.  The Makefile
is still including the changsets in the kernel version so its still
changing the ABI with every commit.

This does work around a problems with kernel versions that don't
include a dash.  The kernel version would not be the same as the RPM
version which meant the kernel headers package would end up
conflicting with itself.