CA-95815: nfs: don't zero out the rest of the page if we hit the EOF on a DIO READ
(commit 67fad106a219e083c91c79695bd1807dde1bf7b9 upstream)

Eryu provided a test program that would segfault when attempting to read
past the EOF on file that was opened O_DIRECT. The buffer given to the
read() call was on the stack, and when he attempted to read past it it
would scribble over the rest of the stack page.

If we hit the end of the file on a DIO READ request, then we don't want
to zero out the rest of the buffer. These aren't pagecache pages after
all, and there's no guarantee that the buffers that were passed in
represent entire pages.