CA-105158: backport netback security and regression fixes from upstream.
This includes a number of security fixes that are present in 3.8 and
3.9-rc1.

And two important regression fixes where netback would drop or
disconnect when recieving previously acceptable packets from the front
end.  Packets with 18 frags (slots) are now accepted again, and it
will no longer signal a fatal error or receiving a packet >= 64 KiB
that was erroneously sent (it will still drop these as this was the
old behaviour).