CP-3791: Sync Xen wallclock and RTC when system time is stepped
When the system time is stepped by (for example) ntpdate, sychronize
the Xen wallclock and the CMOS RTC to the current system time.  This
means that guests that start before NTP syncs and updates the
wallclock.

update_persistent_clock() is called 500 ms before the next second.
Previously, the Xen wallclock would be off by 500 ms.  We handle this
by increasing the precision available to the
x86_platform.get/set_wallclock() calls.