CA-110911: pci: HACK! allow parent parent bus reset if only pciback is bound
When doing a parent bus reset to reset a single function, allow the
reset to occur if the only other devices are either a) not bound; or
b) bound to pciback.

This is hacky as it puts knowledge of pciback into core PCI code.  It
is also risky as the other devices may be in use by other VMs.  We are
assuming the toolstack prevents assigning devices (with no FLR method)
that are behind the same switch port to different VMs.