CA-110911: pciback: add a reset sysfs file if one is missing
If the reset sysfs file is missing from a PCI device, there is no
function reset mechanism.  If all the devices on a bus are co-assigned
to the same domain, we can use a bus reset to reset all the devices.

pciback now creates a reset sysfs file if necessary to do this.

Note: the save/restore of the config space uses the standard
pci_save_state() and pci_restore_state() API which is slightly
different from the save/restore done by the classic kernel's pciback
-- it only saves and restores selected state instead of the whole
configuration state.