CA-111851: fix tboot boot (again!)
Reinstate a variant of the original patch as "x86/xen: during early
setup only 1:1 map the ISA region" does not resolve the issue on 3.6
(but it does with newer kernels).

x86/xen: do not identity map UNUSABLE regions in the machi

If there are UNUSABLE regions in the machine memory map, dom0 will
attempt to map them 1:1 which is not permitted by Xen and the kernel
will crash.

There isn't anything interesting in the UNUSABLE region that the dom0
kernel needs access to so we can avoid making the 1:1 mapping and
treat it as RAM.

We only do this for dom0, as we do not expect any domU to ever have a
UNUSABLE region in their pseudo-physical map.

This fixes a boot failure on hosts with tboot.

tboot marks a region in the e820 map as unusable and the dom0 kernel
would attempt to map this region and Xen does not permit unusable
regions to be mapped by guests.