CA-116499 - Fix kernel spec files to package kernel header files correctly
The update from kernel 3.6 to 3.10 did not update the kernel RPM spec file
to account for the movement of some of the kernel header files which
occurred in 3.8. This commit aligns the kernel spec file more closely with
the Elrepo kernel-ml spec file used for mainline kernel on Centos 6.x.