CA-103351: disable CONFIG_XEN_MCE_LOG in the 64-bit config
Disable CONFIG_XEN_MCE_LOG to remove the mcelog driver.  With this
disabled there is no user of VIRQ_MCA and Xen writes MCE information
to the console instead of queuing them to be read by dom0.

It is expected that this will resolve the MCE related Xen crashes by
avoiding the use of the suspected buggy MCE code.