CA-120950: Call cond_resched() when in the middle of a privcmd hypercall
Hypercalls from privcmd can take a long time.  The hypercall
continuations allow upcalls while the hypercall is in progress so put
a cond_resched() all in the upcall if the upcall interrupted a task
that issues such a hypercall via the privcmd driver.

The should fix the soft lockups in xenguest that may occur with very
long hypercalls when building, saving or restoring large guests.