CA-125706: fix 64-bit kernel boot failure on Nehalem EX
physflat_acpi_madt_oem_check() on Nehalem EX machines results in
physical flat mode being forced on which does not work in a Xen guest.

As a hack, add a test for xen_domain() so physflat isn't forced on.

An upstreamable solution may be to add a Xen specific APIC driver
instead of trying to fiddle with a hardware driver.