CA-114544: Prevent dom0 kernel from touching performance counter MSRs
The performance counters are enabled by default when a K7 or higher AMD
CPU is detected. This patch prevents the devices being probed when
running as a Xen PV domain, since doing so results in warnings to the
Xen console when the behaviour is trapped.

This also improves the message printed on Intel CPUs, which detect
partial hardware support but fail to initialise it. Now both cases
simply report that hardware PMU is not available under Xen.