CP-3759: change kernel ABI and package versions
The kernel (internal) ABI is now 3.10.0+0 where the number after the +
is the ABI version which will be manually adjusted on LCM branches
when the kernel ABI changes.

The package version is 3.10.27 (reflecting the version of the stable
kernel the release is based on).

The package release is the changeset indexes for the patch queue and
base repos.

The kernel package is now named kernel-<KABI-version>.

Packages building kernel modules should use the LINUX_KABI_VERSION
from kernel.inc to get the kernel ABI version.