CA-130758: updated MMIO patch series to v6
This fixes a crash when userspace unmaps MMIO regions.  The M2P access
was faulting because the M2P has holes and then spurious_fault() was
walking the inaccessible page tables for the M2P virtual mapping and
faulting.

This is fixed by making spurious_fault() only walk the tables of
supervisor write or instruction faults.