CA-145482: xen-netback: increase maximum length of internal guest Rx queue
When the length of the internal guest Rx queue was limited to 128
pages (from effectively unlimited) it caused a small (~5%) regression
in aggregate throughput.

Increase the limit to allow a full queue to completely fill the guest
Rx ring.  This increase the amount of packet data that may be queued
per-VIF to 3 MiB (32x 64 KiB packets on the qdisc queue and 1 MiB in
the internal one).