CA-144273: xen-netback: wake queue when guest Rx is ready
netif_carrier_on() was incorrectly assumed to start all the queues.
This meant that if it took 10 s (the drain timeout) for a VIF to go
from CONNECTED to up in the guest, then it was possible for the queue
to be stopped (if > 512 KiB was queued in this time) and the Rx stall
detection to trigger.  When guest Rx became ready, the queue was not
restarted and guest Rx would not work.

Whilst investigating this two more bugs were identified:

- Unsafe use of the unlocked skb_peek(&vif->rx_queue) because skbs
  could be removed concurrently (in the xenwatch thread on
  disconnect). Now only the guest Rx thread removed skbs from the
  internal Rx queue.  This allows the thread to safely use skb_peek()
  without locks.

- The guest Rx wait timeout was not being recalculated when an skb was
  queued to an empty queue.  Now a hand-rolled wait_event_timeout()
  variant is used which recalculates the timeout each iteration.

The queue drain and Rx stall detection is also refactored to to make
the two behaviours independent.

- Queue drain: Only skb's older that the drain timeout are dropped.
  This sets a clear upper bound on the length of time an skb can be
  queued for.  For a guest being destroyed the maximum time needed to
  wait for all the packets it sent to be dropped is still the drain
  timeout (10 s) since it will not be sending new packets.

- Rx stall detection: An Rx stall is detected if there are not enough
  Rx slots and an Rx packet hasn't been queued for a while.  The stall
  timeout defaults to 60 s.  Rx stall/ready only affects the carrier;
  it does not drain, stop or start any queues.