CA-136178 : Xen "mm.c:2830:d0v2 Unknown domain '<domid>'" console spam
Much of this patch is actually related to improving performance of MMAPBATCH.
This work is loosly based on another patch, found here:

http://lists.xen.org/archives/html/xen-devel/2012-12/msg01554.html

This roughly doubles the speed of the privcmod_mmapbatch.
The behaviour is also changed, such that if it encounters the
error ESRCH, it stops and make no further calls to Xen.  Any as yet
unprocessed entries are marked and returned as also failing with ESRCH.