CA-142346: don't IPI all spinlock waiters
Because PV spinlocks re-enabled IRQs in the slow path various
deadlocks could occur (re-enabling IRQs while spinning is not
something bare-metal or the PV ticket locks do).  This was worked
around by IPIing all waiters in the unlock path but this does not
scale well with more VCPUs.

Instead, do not re-enable IRQs in the slow path.  This is a similar to
a fix made on clearwater with the classic ticketlocks (which also
re-enabled IRQs in the slow path).