CP-9934: xen/grant-table: add a mechanism to safely unmap pages that are in use
Introduce gnttab_unmap_refs_async() that can be used to safely unmap
pages that may be in use (ref count > 1).  If the pages are in use the
unmap is deferred and retried later.  This polling is not very clever
but it should be good enough if the cases where the delay is necessary
are rare.

This use case is this for is network storage (block or filesystem
based such as iSCSI or NFS) being used by a block backend using grant
maps (such a blkback or qemu via the gntdev device).

The network storage may complete a block request whilst there is a
queued network packet retry (because the ack from the remote end races
with deciding to queue the retry).

Make blkback and gntdev use this new async grant unmap.  Netback does
/not/ use it because in certain circumstances the zerocopy callback is
called when the frags are still ref'd (e.g., from skb_tx_error() or
netback itself if it needed to copy frags from a frag_list).