CP-9935: Fix get_user_pages() for foreign pages mapped by /dev/xen/gntdev
get_user_pages() works by walking the userspace page tables converting
the PTEs to pages.  For a Xen PV guest the PTE contains a MFN so a
MFN-to-PFN (M2P)lookup is needed before the PFN can be translated into
a page.

If a foreign page is mapped into user space this M2P lookup still
needs to work, but the simple lookup into the M2P would return the PFN
for the foreign guest and not the PFN that page is grant mapped into.

gntdev tried to fix this with an M2P override table which the M2P
lookup could consult for non-local pages.  This does not work because
an MFN can be grant mapped onto more than one local page (P2M is
many-to-one), but the M2P override is only one-to-one.  This can
happen if the foreign guest grants access to the same frame via two or
more grant refs.

get_user_pages() would return the wrong pages in this case, and the
caller would eventually corrupt the VM subsystem state by releasing
references to pages there were in use by something else (or already
free).

Fix this by:

1. Making vm_normal_page() lookup the right page in the new vma->pages
   array (if this is present and the PTE is special).

2. Marking userspace PTEs for grant maps as special.  Either in the
   grant map hypercall if this is supported or by fixing it up
   afterwards.

3. Getting gntdev to set vma->pages.

The M2P override is then redundant and can simply be removed.  As a
consequence we also batch the kernel grant maps and unmaps which
should have a performance benefit.