CP-10765: xen/balloon: fix memory hotplug for PV guests
The Xen balloon driver can hotplug additional memory sections if
needed to meet the current target.  This did not work for PV guests
because the sections must be populated with frames or add_memory()
would fail to build the direct-mapping page tables.

Fix this by populating the memory and updating the p2m before adding
the memory.

The balloon driver would also not select a suitable pseudo-physical
address to place the memory (using just max_pfn), and it could overlap
with other E820 regions causing it to fail.

Add add_memory_resource() to the core memory hotplug sub-system that
accepts an existing "System RAM" resource and make the balloon driver
use allocate_resource() to find a suitably sized and aligned region.

Two p2m fixes are also required.  The additional of p2m_mid_identity
to make large identity regions effeciently introduced a regression in
set_phys_to_machine().  It would not handle setting an entry if it was
using p2m_mid_identity.

Finally, on large >512 GiB hosts, the hotplug region would be placed
beyond the end of the P2M (which is limited to 512 GiB) because all
the (unused) host RAM regions are set as E820_UNUSABLE.  Discarding
these regions instead allows the hotplug region to be within the P2M.