CP-10765: Enable memory hotplug in the Xen balloon driver
Enable XEN_BALLOON_MEMORY_HOTPLUG so additional memory can be added to
guests even if they were started with a lower maximum reservation.
This is particularly useful for increasing the memory in dom0 without
having to adjust the Xen command line and reboot.

Memory is hotplugged by setting the balloon target to above the
initial maximum reservation.

Note:

1. The toolstack will need to increase the maximum.

2. The additional memory is only added in whole sections (128 MiB).

3. The new memory needs to be onlined by writing "online" to memory
   device's "state" node in sysfs.