CA-161069: Ensure build ID in vmlinuz and debuginfo match (fixes systemtap)
systemtap in CentOS 7 checks that the build ID (an ELF note entry) of
the debuginfo matches the running kernel.  As part of extracting the
debuginfo the build ID is modified which means it does not match that
baked into vmlinuz (bzImage).

Add a patch (kbuild-AFTER_LINK.patch) from the Fedora kernel which
adds an AFTER_LINK option to the kernel build.  Use this to fixup the
vmlinux build ID /before/ generating the bzImage.