CA-167157: blktap2: don't get page pools that are being released
blktap_page_pool_get() calls __blktap_kset_find_obj() to get a
reference to a named page pool if it already exists.

This may find a page pool that is in the process of being released but
which has not yet been removed from the kset.  This will cause an oops
when the page pool is used (because, for example, the mempool has been
destroyed).

Additionally, even if we skip kobjects that are being released (by
using something like kobject_get_unless_zero()) the sysfs file may
still exist which would prevent re-adding the page pool with the same
name.

Separate the usage count of the page pool from the underlying kobject
reference count.  When this usage count is zero call kobject_del() to
remove both the sysfs files and the kobject from the kset.  Do this
under the pool_set_mutex to avoid any races with
blktap_page_pool_get_by_name().