CA-166945: Fix Xen's save_fl PVOP to return all flags.
Despite the comment, this PVOP is used by setup_smap() to check whether the
Alignment Check flag is set.

For Xen PV guests, setup_smap() ends up executing a BUG_ON() against an
uninitialised value, and experimentally fails for most 32bit PV guests on
Broadwell hardware.

'pushf' is safe to use in PV guests to get all the flags, but will
unconditionally have the Interrupt Flag set, as the cpu is scheduled in guest
context.  The Interrupt Flag is therefore conditionally masked out if local
events are masked.