CP-11471: Remove 004-xen-balloon-populate-the-hotplug-memory-before- addin.patch and adding xen-hotplug-add-new-memory.patch.
This change enables new memory to be added without prepopulation , one
of the condition for memory to be usable by alloc_xenballoone_pages.
alloc_xenballooned_pages is called by bulk and network backed to
service their request for additional memory. So this patch enables
hotplugged memory to be available to bulk and network backends.