CA-172015: xen-netback: Remove use of fraglist in tx path
If the number of slots used by the packet is XEN_NETBK_LEGACY_SLOTS_MAX,
copy the entire first slot instead of copying part of it and mapping the
rest. This removes the need handle a frag overflow in a separate
fraglist.

This fixes a use-after-free in the OOM error path for the fraglist.