CA-175242: xen-netback: fix crashes on queue init and teardown
Don't leave a stale NAPI instance added to the net device if queue
initialization fails.

Check the allocation for the queues succeeds and raise a fatal error
if it fails.

In case we incorrectly leave a NAPI instance added before freeing a
VIF, don't free the queues until the net device is freed.  This avoids
a use-after-free in free_netdev() when the NAPI instances are deleted.