CA-167633: xen/balloon: Select HOLES_IN_ZONE if hotplug is enabled
Hotplugging memory can result in zones that have holes (and I suspect
that SPARSEMEM can also result in zones with holes).  This can cause a
bogus BUG_ON() to fire if move_freepages() is trying to move a block
of pages that extends into a hole.

HOLES_IN_ZONE was only available for mips and ia64 so it's added to
the generic mm/Kconfig.