CA-173529: Turn off oplocks to disable client side-caching
Oplocks interfere with concurrent access of CIFS files by multiple
clients and hence need to be disabled for CIFS shares to serve as
a shared SR in a XenServer Pool.

This is hopefully a temporary patch till we can figure out a more
elegant way to disable oplocks through an external interface.

Signed-off-by: Chandrika Srinivasan <chandrika.srinivasan@citrix.com>
Reviewed-by: Felipe Franciosi <felipe.franciosi@citrix.com>