CA-164808 - Mask PCIE completion timeout when assigning PCI device to VM
Certain Nouveau driver versions can cause the Nvidia GPU cards to not
respond for a period of time. Any accesses cause a PCIE completion
timeout which can be propogated to host firmware and trigger a NMI
to be injected. Xen treats unexpected NMI as fatal and starts the
crashdump process. Thus the Nouveau driver can triger a host crashdump.

Masking the PCIE completion timeout in PCIE AER register's leads to the
status bit still being set. Host firmware detects this status bit on warm
reset and blocks boot. This behaviour is not acceptable but there is no
reliable mechanism to clear the status bit before a warm reset.

This fix results in the PCIE completion timeout being masked in Intel
specific registers so that the status bit in the PCIE AER registers is
never set. This prevents host firmware from reacting to the issue and
avoids the halt on boot problem.