CA-195850 - Intel P-states not functioning after CPUID leveling work
The CPUID leveling work removed the EST feature bit from PV and HVM guests.

The dom0 kernel used this feature to determine how gather p and c states and
how to setup the ACPI PDC region.

Xen is not capable of activating turbo without the ACPI p-state table.

To fix the regression, force add the EST feature bit if it's been detected
via a native cpuid call, this is the least invasive method without modifying
the Xen hypervisor itself.