XOP-722: xen-pciback: handle buses without bridges when disabling AER
If adding an SR-IOV VF required adding an additional bus, this bus
would be a child of another bus (the parent) but would not itself have
an upstream bridge device.

Attemping to passthrough such a VF would oops when attempting to
disable AER because the root device could not be found.

When finding the root device, walk up the bus tree to find the parent
bus with an upstream bridge device.