CA-204251: x86/iopl/64: Properly context-switch IOPL on Xen PV
[ Commit b7a584598aea7ca73140cb87b40319944dd3393f upstream with static
  cpu feature check replaced with xen_pv_domain(). ]

On Xen PV, regs->flags doesn't reliably reflect IOPL and the
exit-to-userspace code doesn't change IOPL.  We need to context
switch it manually.

I'm doing this without going through paravirt because this is
specific to Xen PV.  After the dust settles, we can merge this with
the 32-bit code, tidy up the iopl syscall implementation, and remove
the set_iopl pvop entirely.

Fixes XSA-171.