CA-206923: xen/setup: Reserve the crashkernel region
Reserve the crashkernel region in the E820. This prevents oustanding DMA
operations overwriting the crashkernel region when the IOMMU is disabled
during kexec. Normally, the IOMMU would not be used for dom0, but it is
used for PV-IOMMU.

This attempts to fix various hangs, memory corruption and segfaults seen
while executing the crashkernel.