CA-211435: Fix host crashes when force-killing a domain using blkback
Wait for pending IO to complete before freeing persistent grants to
avoid a use-after-free when the device is disconnected while IO is
in-flight.

This can be reproduced by running an IO workload in a guest a
force-killing the domain.

Furthermore, don't use persistent grants at all. Persistent grants are
ugly and require a copy in the frontend. Grant mapping performance
should be OK since the locking code has been improved.