CA-211765: Fix IOCTL_PRIVCMD_MMAPBATCH* error handling
When writing error codes into the caller's array, a fastpath was
introduced to skip writing anything until the first error code. However,
it did not correctly increment the pointer into the caller's array, so
the error codes would be copied into the wrong locations in the array if
the index of the first error != 0. To fix, this increment the caller
array pointer in the fastpath.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>