CA-207786: Disable HUGETLBFS
Huge pages don't work in a PV guest and attempting to use them (which
requires mounting /dev/hugetlb) will cause a crash.

This addresses XSA-174 which was not considered to affect XenServer
because exploiting it (by a non-root user) required the host admin to
manually mount the hugetlbfs of /dev/hugetlb as group or other write
permissions, or set vm.hugetlb_shm_group.