CA-215311: xen/balloon: leak pages if their IOMMU mapping cannot be updated
When PV-IOMMU is being used failure to properly update the IOMMU
mappings when returning (freeing) a page from the balloon can result
in memory and storage corruption as DMA to/from this page may be to
the incorrect machine frame.

If the IOMMU mapping fails, the page is leaked as it cannot be safely
used.

Once possible cause of mapping failures is if the BFN (== PFN) already
has a mapping.  This might be possible with hotplugged memory if the
new region is placed over a region that is both 1:1 in the P2M and
used to be RAM in the e820 (so it still has Xen created mappings with
BFN == MFN).  But this combination is not possible since 1:1 regions
are setup before discarding the former RAM regions.

To keep the change the same as the one made for the Linux 4.4, we
clear IOMMU mappings when onlining (as yet unpopulated) page.